Extract Shiny module and extend detection into packages
