+++ crates/oak_db/src/tests.rs
@@ -8,6 +8,7 @@ mod file_imports_at;
 mod file_resolve;
 mod file_resolve_at;
 mod file_root;
+mod file_shiny;
 mod file_source_site;
 mod identifier;
 mod inputs;
