+++ crates/oak_db/src/tests/file_imports.rs
@@ -38,7 +38,7 @@ fn make_installed(db: &mut TestDb, name: &str) -> (crate::Root, Package) {
 
 /// Register a set of installed packages on `LibraryRoots`. Replaces any
 /// previously registered library roots.
-fn install_packages(db: &mut TestDb, names: &[&str]) -> Vec<Package> {
+pub(super) fn install_packages(db: &mut TestDb, names: &[&str]) -> Vec<Package> {
     let mut roots = Vec::new();
     let mut packages = Vec::new();
     for &name in names {
@@ -528,7 +528,7 @@ fn test_testthat_file_includes_package_namespace_imports() {
 
 /// Render `ImportLayer`s to a stable, assertable shape. `File` layers
 /// collapse to their basename.
-fn shape(db: &TestDb, layers: &[ImportLayer]) -> Vec<String> {
+pub(super) fn shape(db: &TestDb, layers: &[ImportLayer]) -> Vec<String> {
     layers
         .iter()
         .map(|layer| match layer {
@@ -1684,387 +1684,4 @@ fn test_tar_source_recurses_into_subdirectories() {
     let mut names: Vec<&str> = pipeline.exports(&db).iter().map(|(name, _)| name).collect();
     names.sort();
     assert_eq!(names, vec!["a_val", "fit_val"]);
-}
-
-/// A workspace of scripts at `ws`, returned with the root and in the order
-/// given.
-fn script_workspace(db: &mut TestDb, scripts: &[(&str, &str)]) -> (crate::Root, Vec<File>) {
-    let root = workspace_root(db, "ws");
-    let files: Vec<File> = scripts
-        .iter()
-        .map(|(path, contents)| {
-            File::new(
-                db,
-                file_path(path),
-                FileRevision::zero(),
-                Some(contents.to_string()),
-                None,
-            )
-        })
-        .collect();
-    root.set_scripts(db).to(files.clone());
-    db.workspace_roots().set_roots(db).to(vec![root]);
-    (root, files)
-}
-
-#[test]
-fn test_shiny_entry_sees_global_then_the_autoloaded_directory() {
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/global.R", "cfg <- 1\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-        ("ws/R/b.R", "b_val <- 2\n"),
-    ]);
-
-    // LIFO: the `R/` files run last and into an inner environment, so they
-    // shadow `global.R`.
-    assert_eq!(shape(&db, files[0].imports(&db)), vec![
-        "File(b.R)".to_string(),
-        "File(a.R)".to_string(),
-        "File(global.R)".to_string(),
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_shiny_autoload_does_not_recurse() {
-    // `loadSupport()` lists `R/` with `recursive=FALSE`, unlike `tar_source()`.
-    let mut db = TestDb::new();
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-        ("ws/R/models/fit.R", "fit_val <- 2\n"),
-    ]);
-
-    assert_eq!(shape(&db, files[0].imports(&db)), vec![
-        "File(a.R)".to_string()
-    ]);
-}
-
-#[test]
-fn test_shiny_ui_and_server_do_not_see_each_other() {
-    // Shiny evaluates them into sibling environments, so neither is part of the
-    // other's support set. Pinned so nobody later "fixes" it.
-    let mut db = TestDb::new();
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/ui.R", "shinyUI(fluidPage())\n"),
-        ("ws/server.R", "shinyServer(function(input, output) NULL)\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-    ]);
-
-    assert_eq!(shape(&db, files[0].imports(&db)), vec![
-        "File(a.R)".to_string()
-    ]);
-    assert_eq!(shape(&db, files[1].imports(&db)), vec![
-        "File(a.R)".to_string()
-    ]);
-}
-
-#[test]
-fn test_shiny_disable_autoload_drops_the_directory_but_keeps_global() {
-    // The opt-out returns from `loadSupport()` after the `global.R` block.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/global.R", "cfg <- 1\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-        ("ws/R/_disable_autoload.R", "\n"),
-    ]);
-    let a = files[2];
-
-    assert_eq!(shape(&db, files[0].imports(&db)), vec![
-        "File(global.R)".to_string(),
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-
-    // `a.R` itself is never sourced by `loadSupport()` when disabled, so it
-    // gets none of the Shiny treatment either: no inherited `global.R`, and
-    // no implicit `shiny` attach. Plain `R/` collation applies instead --
-    // `_disable_autoload.R` is a real file in the directory, so it shows up
-    // as an ordinary (empty) collation sibling.
-    assert_eq!(shape(&db, a.imports(&db)), vec![
-        "File(_disable_autoload.R)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_confirmed_entry_point_with_nothing_to_inherit_still_attaches_shiny() {
-    // `loadSupport()` calls `require(shiny)` unconditionally, before it even
-    // checks whether `global.R` or `R/` exist. So a lone `app.R` with the
-    // marker call has no `File` layer to inherit, but it's still a real Shiny
-    // app and still gets the attach.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[("ws/app.R", "shinyApp(ui, server)\n")]);
-
-    assert_eq!(shape(&db, files[0].imports(&db)), vec![
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_autoloaded_file_with_no_global_still_gets_the_implicit_shiny_attach() {
-    // `shiny_autoload` returns `Some(vec![])` here, not `None`: no `global.R`
-    // to inherit from, but `loadSupport()` still ran `require(shiny)` first.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-        ("ws/R/b.R", "b_val <- 2\n"),
-    ]);
-
-    assert_eq!(shape(&db, files[2].imports(&db)), vec![
-        "File(a.R)".to_string(),
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_autoloaded_file_sees_global_and_the_implicit_shiny_attach() {
-    // `loadSupport()` evaluates `global.R` first and opens with
-    // `require(shiny)`, so a module file gets both without asking.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "dplyr", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/global.R", "library(dplyr)\n"),
-        ("ws/R/mod.R", "mod_val <- 1\n"),
-        ("ws/R/util.R", "util_val <- 2\n"),
-    ]);
-
-    assert_eq!(shape(&db, files[2].imports(&db)), vec![
-        "File(util.R)".to_string(),
-        "File(global.R)".to_string(),
-        "Package(dplyr)".to_string(),
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_r_directory_without_an_entry_point_ignores_global() {
-    // A `global.R` beside an `R/` only means something to Shiny, so without an
-    // entry point the directory keeps the plain collation view.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/global.R", "cfg <- 1\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-        ("ws/R/b.R", "b_val <- 2\n"),
-    ]);
-
-    assert_eq!(shape(&db, files[1].imports(&db)), vec![
-        "File(b.R)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_app_r_without_the_shiny_app_call_is_not_detected() {
-    // Same filename and layout as a real entry point, but the content marker
-    // `getShinyFileType()` requires (`shinyApp` for `app.R`) is missing, so it
-    // reads as a plain script that happens to be named `app.R`.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "1 + 1\n"),
-        ("ws/global.R", "cfg <- 1\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-    ]);
-
-    assert_eq!(shape(&db, files[0].imports(&db)), vec![
-        "Package(base)".to_string()
-    ]);
-}
-
-#[test]
-fn test_shiny_layers_backdate_on_unrelated_script_change() {
-    // `shiny_autoload` reads every workspace root's scripts, so a script added
-    // anywhere re-executes it. Filtered to the app directory the answer is
-    // unchanged, so salsa backdates it and `cross_file_layers` stands. Reading
-    // the scripts straight from `cross_file_layers` would lose that.
-    let mut db = TestDb::new();
-    let (root, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-    ]);
-    let app = files[0];
-
-    let _ = app.imports(&db);
-    assert_eq!(db.executions("cross_file_layers"), 1);
-
-    let elsewhere = File::new(
-        &db,
-        file_path("ws/other/z.R"),
-        FileRevision::zero(),
-        Some("z_val <- 1\n".to_string()),
-        None,
-    );
-    let mut scripts = files.clone();
-    scripts.push(elsewhere);
-    root.set_scripts(&mut db).to(scripts);
-
-    let _ = app.imports(&db);
-    assert_eq!(db.executions("cross_file_layers"), 1);
-}
-
-#[test]
-fn test_editing_entry_file_body_reexecutes_only_the_classifier() {
-    // `is_shiny_entry_file` is its own tracked query specifically so its
-    // content read backdates independently. Without it, every `R/` file in
-    // the app would reach `app.R`'s `source_text` through `is_shiny_dir`, and
-    // any edit there -- marker-preserving or not -- would re-execute their
-    // `shiny_autoload`. With it, only the classifier re-runs; its boolean is
-    // unchanged, so `shiny_autoload` and `cross_file_layers` stay cached.
-    let mut db = TestDb::new();
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-    ]);
-    let (app, a) = (files[0], files[1]);
-
-    let _ = app.imports(&db);
-    let _ = a.imports(&db);
-    let classifier_before = db.executions("is_shiny_entry_file");
-    let autoload_before = db.executions("shiny_autoload");
-    let cross_file_layers_before = db.executions("cross_file_layers");
-
-    // A comment above the call: `shinyApp` stays in the text.
-    app.set_source_text_override(&mut db).to(Some(
-        "# a harmless comment\nshinyApp(ui, server)\n".to_string(),
-    ));
-
-    let _ = app.imports(&db);
-    let _ = a.imports(&db);
-
-    assert!(db.executions("is_shiny_entry_file") > classifier_before);
-    assert_eq!(db.executions("shiny_autoload"), autoload_before);
-    assert_eq!(db.executions("cross_file_layers"), cross_file_layers_before);
-}
-
-#[test]
-fn test_removing_the_shiny_app_call_drops_the_shiny_layers() {
-    // The other side of the firewall: this is a real invalidation, not
-    // churn. Once `app.R`'s content no longer classifies it as an entry
-    // point, both the app and its `R/` sibling lose the `shiny` layer.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "shinyApp(ui, server)\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-    ]);
-    let (app, a) = (files[0], files[1]);
-
-    assert_eq!(shape(&db, app.imports(&db)), vec![
-        "File(a.R)".to_string(),
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-    assert_eq!(shape(&db, a.imports(&db)), vec![
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-
-    app.set_source_text_override(&mut db)
-        .to(Some("1 + 1\n".to_string()));
-
-    assert_eq!(shape(&db, app.imports(&db)), vec![
-        "Package(base)".to_string()
-    ]);
-    assert_eq!(
-        shape(&db, a.imports(&db)),
-        vec!["Package(base)".to_string()]
-    );
-}
-
-#[test]
-fn test_adding_the_shiny_app_call_creates_the_shiny_layers() {
-    // The remaining direction: a plain script named `app.R` picks up the
-    // layers the moment an edit gives it the marker call, on both itself and
-    // its `R/` sibling.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let (_, files) = script_workspace(&mut db, &[
-        ("ws/app.R", "1 + 1\n"),
-        ("ws/R/a.R", "a_val <- 1\n"),
-    ]);
-    let (app, a) = (files[0], files[1]);
-
-    assert_eq!(shape(&db, app.imports(&db)), vec![
-        "Package(base)".to_string()
-    ]);
-    assert_eq!(
-        shape(&db, a.imports(&db)),
-        vec!["Package(base)".to_string()]
-    );
-
-    app.set_source_text_override(&mut db)
-        .to(Some("shinyApp(ui, server)\n".to_string()));
-
-    assert_eq!(shape(&db, app.imports(&db)), vec![
-        "File(a.R)".to_string(),
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-    assert_eq!(shape(&db, a.imports(&db)), vec![
-        "Package(shiny)".to_string(),
-        "Package(base)".to_string(),
-    ]);
-}
-
-#[test]
-fn test_package_contained_shiny_app_is_out_of_scope() {
-    // golem-style `inst/app/app.R`: a real Shiny app, but owned by a package.
-    // `cross_file_layers`'s Shiny arms are gated on `package(db) == None`, so
-    // a file with a package back-pointer never reaches them, regardless of
-    // its own name or content -- this stays standalone rather than crashing
-    // or guessing. Explicitly out of scope for now: even lifting the `None`
-    // gate wouldn't be enough, since `files_in_directory` only lists
-    // `workspace_scripts`, never `Package::scripts`, so `app.R`'s own `R/`
-    // siblings under `inst/app/` wouldn't be found either. See the "Shiny-in-
-    // a-package" gap in the design note.
-    let mut db = TestDb::new();
-    install_packages(&mut db, &["base", "shiny"]);
-    let workspace = workspace_root(&db, "w");
-    let pkg = Package::new(
-        &db,
-        file_path("w/pkg/DESCRIPTION"),
-        "pkg".to_string(),
-        FileRevision::zero(),
-        FileRevision::zero(),
-        None,
-        None,
-        Vec::new(),
-        Vec::new(),
-    );
-    let app = File::new(
-        &db,
-        file_path("w/pkg/inst/app/app.R"),
-        FileRevision::zero(),
-        Some("shinyApp(ui, server)\n".to_string()),
-        Some(pkg),
-    );
-    let helper = File::new(
-        &db,
-        file_path("w/pkg/inst/app/R/mod.R"),
-        FileRevision::zero(),
-        Some("mod_val <- 1\n".to_string()),
-        Some(pkg),
-    );
-    pkg.set_scripts(&mut db).to(vec![app, helper]);
-    workspace.set_packages(&mut db).to(vec![pkg]);
-    db.workspace_roots().set_roots(&mut db).to(vec![workspace]);
-
-    assert_eq!(shape(&db, app.imports(&db)), vec![
-        "Package(base)".to_string()
-    ]);
 }
